feat: auth api 및 type 구현
