+++ app/_types/auth.type.ts
@@ -1,0 +1,11 @@
+interface LoginUserParams {
+  email: string;
+  password: string;
+}
+
+interface UpdatePasswordParams {
+  password: string;
+  newPassword: string;
+}
+
+export type { LoginUserParams, UpdatePasswordParams };
